update: switch from double to float for coordinates

+++ include/DEM.hpp
@@ -9,25 +9,25 @@
 class DEM {
 private:
     struct Index {
-        double row;
-        double column;
+        float row;
+        float column;
     };
 
     int read(const std::string& filepath);
-    Index index(double latitude, double longitude);
+    Index index(float latitude, float longitude);
 
 
 public:
     struct Coordinate {
-        double latitude;
-        double longitude;
+        float latitude;
+        float longitude;
 
         Coordinate() {
             this->latitude = 0;
             this->longitude = 0;
         };
 
-        Coordinate(double latitude, double longitude) {
+        Coordinate(float latitude, float longitude) {
             if (latitude > 90 || latitude < -90 || longitude > 180 || longitude < -180) {
                 std::string e = "invalid coordinates (" + std::to_string(latitude) +  ":" +  std::to_string(longitude) + ")";
                 throw std::runtime_error(e);
@@ -58,9 +58,9 @@ public:
     struct Type {
         unsigned int nrows;         // no. of DEM values available in row
         unsigned int ncols;         // no. of DEM values available in column
-        double yllcorner;           // bottom left latitude
-        double xllcorner;           // bottom left longitude
-        double cellsize;            // distance (in radians) between every DEM values
+        float yllcorner;           // bottom left latitude
+        float xllcorner;           // bottom left longitude
+        float cellsize;            // distance (in radians) between every DEM values
         short int nodata;           // invalid DEM value representation
 
         Type() {
@@ -72,7 +72,7 @@ public:
             this->nodata = 0;
         };
 
-        Type (unsigned int nrows, unsigned int ncols, double yllcorner, double xllcorner, double cellsize, short int nodata) {
+        Type (unsigned int nrows, unsigned int ncols, float yllcorner, float xllcorner, float cellsize, short int nodata) {
             if (yllcorner > 90 || yllcorner < -90 || xllcorner > 180 || xllcorner < -180) {
                 std::string e = "invalid coordinates (" + std::to_string(yllcorner) +  ":" +  std::to_string(xllcorner) + ")";
                 throw std::runtime_error(e);
@@ -94,7 +94,7 @@ public:
     DEM(const Type& type, const std::string& filepath);
     ~DEM();
 
-    bool check_coordinates_bounds(double latitude, double longitude);
-    short int altitude(double latitude, double longitude);
-    double interpolated_altitude(double latitude, double longitude);
+    bool check_coordinates_bounds(float latitude, float longitude);
+    short int altitude(float latitude, float longitude);
+    float interpolated_altitude(float latitude, float longitude);
 };
